validation middlewares fixed. request route added

+++ backend/src/controllers/user.controller.js
@@ -7,7 +7,7 @@ import awsFileGet from "../utils/awsFileHandle/awsFileGet.js"
 export const createUser = async (req, res) => {
     try {
         const userExists = await User.findOne({ email: req.body.email });
-        if (userExists) throw new Error("El email ya se encuentran registrados en la base de datos")
+        if (userExists) throw new Error("El email ya se encuentra registrado en la base de datos")
 
         const userProps = Object.assign({}, req.body)
         userProps.password = await User.encryptPassword(userProps.password)
@@ -19,7 +19,7 @@ export const createUser = async (req, res) => {
 
     } catch (err) {
         console.log(err)
-        res.status(500).json(err)
+        res.status(500).json(err.message)
     }
 }
 
@@ -27,14 +27,14 @@ export const loginUser = async (req, res) => {
     try {
         const { email, password } = req.body
 
-        const user = await User.findOne({ email })
+        const user = await User.findOne({ email }).select({ email: 1, password: 1 })
         if (!user) throw new Error("El usuario o contraseña son incorrectos")
 
         const passwordMatch = await User.comparePassword(password, user.password);
         if (!passwordMatch) throw new Error("El usuario o contraseña son incorrectos")
 
         const userData = {
-            id: user._id,
+            id: user._id.toString(),
             email: user.email,
             fullName: user.fullName
         }
@@ -43,18 +43,18 @@ export const loginUser = async (req, res) => {
 
         return res.status(200).header("token", dataToken).json({ message: "Login Succesfully", userData })
     } catch (err) {
-        res.status(500).json(err);
+        res.status(500).json(err.message)
     }
 }
 
 export const getUserbyId = async (req, res) => {
     try {
         const user = await User.findById(req.params.id).lean()
-        user.img = await awsFileGet(user.img)
+        if (user.img) user.img = await awsFileGet(user.img)
 
         res.status(200).json(user)
     } catch (err) {
-        res.status(500).json(err);
+        res.status(500).json(err.message)
     }
 }
 
@@ -68,7 +68,7 @@ export const addImg = async (req, res) => {
 
         res.status(200).json(user)
     } catch (err) {
-        res.status(500).json(err);
+        res.status(500).json(err.message)
     }
 }
 
@@ -78,21 +78,23 @@ export const editUser = async (req, res) => {
         res.status(200).json(user)
 
     } catch (err) {
-        res.status(500).json(err);
+        res.status(500).json(err.message)
     }
 }
 
 export const deleteUser = async (req, res) => {
     try {
-        const user = await User.findById(req.params.id)
+        const user = await User.findById(req.params.id).select({ password: 1 })
+        console.log(req.body.password)
+        if (!req.body.password) throw new Error("Ingrese su contraseña para confirmar el borrado de la cuenta")
         const passwordMatch = await User.comparePassword(req.body.password, user.password);
         if (!passwordMatch) throw new Error("La contraseña es incorrecta")
 
-        await awsFileDelete(user.img)
+        if (user.img) await awsFileDelete(user.img)
         await User.findByIdAndRemove(req.params.id)
-        res.status(200).json(user)
+        res.status(200).json({ message: "Usuario eliminado correctamente", id: user._id })
 
     } catch (err) {
-        res.status(500).json(err);
+        res.status(500).json(err.message)
     }
 }
